GameEvent Elements
GameEvents are Scriptable Objects that act as a decoupled delegate.

they can currently take a payload of
- string
- float
- bool
- int
- enum GameEventPayLoad.States

Mainly used in conjunction with EventListener, that will allows actions to be designed in the GUI interface.

Eventraiser is a script meant mostly for manual testing. It is currently missing its Editor override script to create its main button functionality. so pretty useless in its current state.