Adding Major Bug fixes, Added visual responses
Prepared Player response to show player group that was correct. Currently the system is using the song index id as a group id. the song index id is always unique to the group.

Fixed a great deal of race conditions that occured durring the change over to singleton controller behaviour.

Fixed Race conditions that occured in the lobby view by giving the lobby its own manager that is always awake. this allows the lobby to properly hook onto the various controller singleton instances. Start is only called on enabled gameobjects, and Awake method often jumbles the order between scripts. Solution: to move controller to an active gameobject, and allow start to be called after awake method has instanciated the singleton controllers.

Fixed an issue with changing genre before another client has joined the game. The syncvars work great, but the hooks are note fired with the syncvar in initialization. this is anoying. Fixed by assigning the newly placed lobby view manager a listener when new players are added, to fire the genre updated method. Note: in the future, if the game gains complexity, additional variables will most likely need to be handled this way.

Encoutered Fatal bugs when converting the game manager into a singleton.
- The primary issue, the singleton methodoly breaks completely when set into a NetworkBehaviour. The editor will work perfectly fine, as intended, but any builds result in a complete reference loss, and will not function. This has been confirmed by the community.

- current solution : leave the GameController as a NetworkBehaviour only. Attached the GameController as an extension of the MainController and reference the GameController as a method of the MainController instance. This is not ideal, but I weighed the options at hand, and attempted a few other work arounds, this was by far, the most code friendly and readable method I could come up with.

Added in visual popups that display app conditions
- new error  notice that displays you have been disconnected, and the error reason.
- new message display that informs of various games states, such as Searching for game, connecting to game, and making new game. these states are currently associated with the auto join feature.